+++ src/Mumsys_DateTime_Exception.php
@@ -1,0 +1,29 @@
+<?php
+
+/**
+ * Mumsys_DateTime_Exception
+ * for MUMSYS Library for Multi User Management System (MUMSYS)
+ *
+ * @license LGPL Version 3 http://www.gnu.org/licenses/lgpl-3.0.txt
+ * @copyright Copyright (c) by Florian Blasel, 2021
+ *
+ * @category    Mumsys
+ * @package     Library
+ * @subpackage  DateTime
+ * @version 1.0.0
+ * Created: 2021-12-17
+ */
+
+
+/**
+ * Generic exception thrown by DateTime classes.
+ *
+ * @category    Mumsys
+ * @package     Library
+ * @subpackage  DateTime
+ */
+class Mumsys_DateTime_Exception
+    extends Mumsys_Exception
+{
+
+}