+++ tests/src/Mumsys_Variable_Manager_DefaultTest.php
@@ -68,7 +68,7 @@ class Mumsys_Variable_Manager_DefaultTest
      *  Version ID of Mumsys_Variable_Manager_Default
      * @var string
      */
-    protected $_version = '2.3.7';
+    protected $_version = '2.3.9';
 
     /**
      * @var array
@@ -165,6 +165,22 @@ class Mumsys_Variable_Manager_DefaultTest
 
         $item->setValue( '' );
         $actual2 = $this->_object->validate();
+
+        $this->assertingTrue( $actual1 );
+        $this->assertingFalse( $actual2 );
+    }
+
+
+    /**
+     * @covers Mumsys_Variable_Manager_Default::validateAll
+     */
+    public function testValidateAll()
+    {
+        $item = $this->_object->getItem( 'username' );
+        $actual1 = $this->_object->validateAll();
+
+        $item->setValue( '' );
+        $actual2 = $this->_object->validateAll();
 
         $this->assertingTrue( $actual1 );
         $this->assertingFalse( $actual2 );
@@ -334,12 +350,22 @@ class Mumsys_Variable_Manager_DefaultTest
 
         $actualBA = $this->_object->validateRegex( $itemB );
 
+        // c integer not possible in regex (preg try catch) 4CC
+        $itemC = clone $itemA;
+        $itemC->setType( 'string' );
+        $itemC->setValue( 12345 );
+        $actualCC = $this->_object->validateRegex( $itemA );
+
+        // comparison
+
         //A
         $this->assertingTrue( $actualAA );
         $this->assertingFalse( $actualAB );
         $this->assertingFalse( $actualAC );
         //B
         $this->assertingTrue( $actualBA );
+        //C
+        $this->assertingFalse( $actualCC );
     }